requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data visualization
matplotlib>=3.7.0
plotly>=5.14.0
seaborn>=0.12.0

# Technical analysis
ta-lib>=0.4.0
ta>=0.11.0

# Sentiment analysis
transformers>=4.30.0
torch>=2.0.0
nltk>=3.8.0
textblob>=0.17.0
vaderSentiment>=3.3.2

# News scraping
selenium>=4.10.0
newspaper3k>=0.2.8
feedparser>=6.0.10

# API clients
yfinance>=0.2.0

# Database
sqlite3-python>=1.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
colorama>=0.4.6
tabulate>=0.9.0
openpyxl>=3.1.0

# Date handling
python-dateutil>=2.8.2
